requirements.txt
asgiref==3.2.10
Django==3.0.9
django-modeltranslation==0.15.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
